loadbalancer: Move tests to their own package
Move the script tests under new tests package to make sure the core
package is clean and does not have many dependencies.
