+++ pkg/loadbalancer/stress.sh
@@ -2,7 +2,7 @@
 set -eux
 [ ! -f $HOME/go/bin/stress ] && go install golang.org/x/tools/cmd/stress@latest
 
-DIRS=". redirectpolicy"
+DIRS="tests experimental/redirectpolicy"
 
 for dir in $DIRS; do
   pushd $dir
