+++ pkg/loadbalancer/tests/script_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package experimental_test
+package tests
 
 import (
 	"context"
