+++ pkg/loadbalancer/tests/testdata/external-clusterip.txtar
