+++ pkg/loadbalancer/tests/testdata/pruning.txtar
