+++ pkg/loadbalancer/tests/testdata/svc-type-annotation.txtar
